+++ src/term.c
@@ -2039,7 +2039,6 @@ match_emoji(termchar * d, int maxlen)
 static void
 emoji_show(int x, int y, struct emoji e, int elen, cattr eattr, ushort lattr)
 {
-  (void)eattr;
   wchar * efn;
   void * * bufpoi;
   int * buflen;
@@ -2056,6 +2055,11 @@ emoji_show(int x, int y, struct emoji e, int elen, cattr eattr, ushort lattr)
 #ifdef debug_emojis
   printf("emoji_show @%d:%d..%d it %d seq %d idx %d <%ls>\n", y, x, elen, !!(eattr.attr & ATTR_ITALIC), e.seq, e.idx, efn);
 #endif
+
+  // Emoji overhang
+  if (elen == 1 && (eattr.attr & TATTR_OVERHANG))
+    elen = 2;
+  //printf("emoj @%d:%d len %d\n", y, x, elen);
   if (efn && *efn)
     win_emoji_show(x, y, efn, bufpoi, buflen, elen, lattr, eattr.attr & ATTR_ITALIC);
 }
@@ -2122,6 +2126,14 @@ term_paint(void)
 
    /*
     * Pre-loop: identify emojis and emoji sequences.
+      This is a hacky approach to handling emoji sequences; problems:
+      - temporary attributes are written back into original char matrix 
+        rather than newchars, yielding:
+      - if the emoji sequences gets broken later by partial overwriting,
+        rendering will become inconsistent
+      - rendering of emojis within bidi lines is only partially correct
+      Moving emoji sequences handling into the first loop below, on the 
+      other hand, would break it even more in bidi lines.
     */
     // Prevent nested emoji sequence matching from matching partial subseqs
     int emoji_col = 0;  // column from which to match for emoji sequences
@@ -2166,6 +2178,16 @@ term_paint(void)
 
           // modify character data to trigger later emoji display
           if (ok && equalattrs) {
+            // Emoji overhang
+            if (e.len == 1 && j + 1 < term.cols
+             && iswspace(d[1].chr) && !d[1].cc_next
+             && d[1].chr != 0x1680 && d[1].chr != 0x3000
+             && d->attr.attr == d[1].attr.attr)
+            {
+              d->attr.attr |= TATTR_OVERHANG;
+              // also mark adjacent space to suppress its display
+              d[1].attr.attr |= TATTR_OVERHANG;
+            }
             d->attr.attr &= ~ATTR_FGMASK;
             d->attr.attr |= TATTR_EMOJI | e.len;
 
@@ -2182,6 +2204,7 @@ term_paint(void)
             }
             else
               tattr = d->attr;
+
             // inhibit rendering of subsequent emoji sequence components
             for (int i = 1; i < e.len; i++) {
               d[i].attr.attr &= ~ATTR_FGMASK;
@@ -2190,6 +2213,10 @@ term_paint(void)
             }
           }
         }
+#ifdef handle_symbol_overhang_here
+        else {  // not an emoji
+        }
+#endif
       }
 
       d->attr = tattr;
@@ -2374,7 +2401,11 @@ term_paint(void)
       * Check the font we'll _probably_ be using to see if
       * the character is wide when we don't want it to be.
       */
-      if (tchar >= 0xE0B0 && tchar < 0xE0C0) {
+      // Emoji overhang
+      if (tattr.attr & TATTR_OVERHANG) {
+        // don't tamper with width of overhanging characters
+      }
+      else if (tchar >= 0xE0B0 && tchar < 0xE0C0) {
         // special handling for geometric "Powerline" symbols
         tattr.attr |= TATTR_ZOOMFULL;
         if (cs_ambig_wide) {
@@ -2382,7 +2413,11 @@ term_paint(void)
         }
       }
 #ifdef ignore_private_use_for_auto_narrowing
-      else if (tchar >= 0xE000 && tchar < 0xF900) {
+#warning auto-narrowing exemption for private use now handled by Symbol overhang below
+      else if ((tchar >= 0xE000 && tchar < 0xF900)
+            || (tchar >= 0xDB80 && tchar < 0xDC00)
+              )
+      {
         // don't tamper with width of Private Use characters
       }
 #endif
@@ -2399,6 +2434,7 @@ term_paint(void)
         }
 
         if ((tattr.attr & TATTR_WIDE) == 0
+            && cfg.char_narrowing < 100
             && win_char_width(xch, tattr.attr) == 2
             // && !(line->lattr & LATTR_MODE) ? "do not tamper with graphics"
             // && is_ambigwide(tchar) ? but then they will be clipped...
@@ -2406,22 +2442,25 @@ term_paint(void)
         {
           //printf("[%d:%d] narrow? %04X..%04X\n", i, j, tchar, chars[j + 1].chr);
           if (
-              // do not narrow various symbol ranges
+              // do not narrow various symbol ranges;
+              // this is a bit redundant with Symbol overhang below
                  (xch >= 0x2190 && xch <= 0x25FF)
               || (xch >= 0x27C0 && xch <= 0x2BFF)
              )
           {
             //tattr.attr |= TATTR_NARROW1; // ?
           }
-          else
+          else {
 #ifdef failed_attempt_to_tame_narrowing
+#warning this is now handled for specific character ranges by Symbol overhang below
             if (j + 1 < term.cols && chars[j + 1].chr != ' ')
 #endif
             tattr.attr |= TATTR_NARROW;
             //if (ch != 0x25CC)
             //printf("char %lc U+%04X narrow %d ambig %d\n", xch, xch, !!(tattr.attr & TATTR_NARROW), is_ambigwide(xch));
+          }
         }
-        else if (tattr.attr & TATTR_WIDE
+        else if ((tattr.attr & TATTR_WIDE)
                  // guard character expanding properly to avoid 
                  // false hits as reported for CJK in #570,
                  // considering that Windows may report width 1 
@@ -2447,6 +2486,34 @@ term_paint(void)
       }
       else if (dispchars[j].attr.attr & TATTR_NARROW) {
         tattr.attr |= TATTR_NARROW;
+      }
+
+     /* Symbol overhang */
+//#define IGNOVRHANG (FONTFAM_MASK | TATTR_WIDE)
+// we cannot support overhang over double-width space (TATTR_WIDE);
+// that would produce artefacts in the scrollback;
+#define IGNOVRHANG (FONTFAM_MASK)
+      if (tchar >= 0x0900 && j + 1 < term.cols
+       && iswspace(chars[j + 1].chr) && !chars[j + 1].cc_next
+       && chars[j + 1].chr != 0x1680
+       && (chars[j].attr.attr & ~IGNOVRHANG) == (chars[j + 1].attr.attr & ~IGNOVRHANG)
+         )
+      {
+        //printf("symb @%d:%d overhang %d attr %08llX attr1 %08llX\n", i, j, !!(chars[j].attr.attr & TATTR_OVERHANG), chars[j].attr.attr, chars[j + 1].attr.attr);
+        if (
+             (tchar >= 0x20A0 && tchar < 0x2400)  // Symbols
+          || (tchar >= 0x2460 && tchar < 0x2500)  // Symbols
+          || (tchar >= 0x25A0 && tchar < 0x2800)  // Symbols
+          || (tchar >= 0x2900 && tchar < 0x2C00)  // Symbols
+          || (tchar >= 0xE000 && tchar < 0xF900)  // Private Use
+          || (tchar >= 0xDB80 && tchar <= 0xDBFF)  // Private Use
+          || (tchar >= 0xDB3C && tchar <= 0xD83E)  // Symbols
+          || indicwide(tchar) || extrawide(tchar)
+           )
+        {
+          tattr.attr |= TATTR_OVERHANG;
+          tattr.attr &= ~TATTR_NARROW;
+        }
       }
 
 #define dont_debug_width_scaling
@@ -2614,7 +2681,7 @@ term_paint(void)
       if (dispchars[j].attr.attr & DATTR_STARTRUN) {
         laststart = j;
         dirtyrect = false;
-        if (firstitalicstart < 0 && newchars[j].attr.attr & ATTR_ITALIC)
+        if (firstitalicstart < 0 && newchars[j].attr.attr & (ATTR_ITALIC | TATTR_OVERHANG))
           firstitalicstart = j;
       }
 
@@ -2637,7 +2704,7 @@ term_paint(void)
         dirtyrect = true;
         prevdirtyitalic = false;
       }
-      if (dirtyrect && dispchars[j].attr.attr & ATTR_ITALIC)
+      if (dirtyrect && dispchars[j].attr.attr & (ATTR_ITALIC | TATTR_OVERHANG))
         prevdirtyitalic = true;
       else if (dispchars[j].attr.attr & DATTR_STARTRUN)
         prevdirtyitalic = false;
@@ -2782,6 +2849,10 @@ term_paint(void)
               eattr.attr &= ~ATTR_REVERSE;
             }
 
+            // Emoji overhang
+            if (elen == 1 && attr.attr & TATTR_OVERHANG)
+              elen = 2;
+            // fill emoji background
             win_text(x, y, esp, elen, eattr, textattr, lattr, has_rtl, false, 1);
             flush_text();
           }
@@ -2809,10 +2880,19 @@ term_paint(void)
         }
 #endif
       }
+      else if ((attr.attr & TATTR_OVERHANG) && iswspace(*text)
+             // skip the skipping if overhanging char was meanwhile changed
+             && start && (newchars[start - 1].attr.attr & TATTR_OVERHANG)
+              )
+      {
+        // Emoji overhang
+        // do not output adjacent space after overhanging emoji;
+        return;
+      }
       else if (overlaying) {
         return;
       }
-      else if (attr.attr & (ATTR_ITALIC | TATTR_COMBDOUBL)) {
+      else if (attr.attr & (ATTR_ITALIC | TATTR_COMBDOUBL | TATTR_OVERHANG)) {
         win_text(x, y, text, len, attr, textattr, lattr, has_rtl, false, 1);
         flush_text();
         ovl_x = x;